Extracted image modal into component and place it into a popup folder with rating-summary component  and simplified full view icon
