+++ frontend/src/popups/RatingSummary.jsx
